Move error enums to their own files
